Draft impl of functional API

+++ src/responses/login_info.rs
@@ -9,13 +9,13 @@ use crate::con_opts::ProtocolVersion;
 #[serde(rename_all = "camelCase")]
 pub struct LoginInfo {
     protocol_version: ProtocolVersion,
-    session_id: u32,
+    session_id: u64,
     release_version: String,
     database_name: String,
     product_name: String,
-    max_data_message_size: u32,
-    max_identifier_length: u16,
-    max_varchar_length: u16,
+    max_data_message_size: u64,
+    max_identifier_length: u64,
+    max_varchar_length: u64,
     identifier_quote_string: String,
     time_zone: String,
     time_zone_behavior: String,
